moved sanitisation types to utils/setlist
